chore: Move referral to a separtate app.
