+++ products/referrals/backend/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class ReferralsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "products.referrals.backend"
+    label = "referrals"
